Move webgl-specific shared code into addon
Never got moved when canvas addon was removed
