+++ addons/addon-webgl/src/CellColorResolver.ts
